// darkhorse-dotnet/toolkit/packages/gantt/src/components/GanttSplitter.d.ts
/**
 * GanttSplitter - Resizable splitter between grid and timeline
 */
import React from 'react';
export interface GanttSplitterProps {
    position: number;
    onResize: (position: number) => void;
}
export declare const GanttSplitter: React.FC<GanttSplitterProps>;
//# sourceMappingURL=GanttSplitter.d.ts.map